[AMDGPU][GFX940] Simplify definitions of GLOBAL_ATOMIC
This change aims to remove unnecessary SubtargetPredicate override
for Flat_Global_Atomics for gfx940.

In current implementation we override SubtargetPredicate twice.
First time when creating definition for Pseudo Instruction in which
SubtargetPredicate is changed to 'isGFX90APlus' and later when
creating Real instruction SubtargetPredicate is changed to 'isGFX940Plus'.
This should not be needed because multiclass 'FLAT_Global_Atomic_Pseudo'
sets up SubtargetPredicate to 'HasFlatGlobalInsts' and all ASICs Generations
should add this feature if they support it.

Additionaly FLAT_Global_Real_Atomics_gfx940 sets up AssemblerPredicate to
'isGFX940Plus' which later is added to final list of all Predicates
for Real instructions in 'list<Predicate> Predicates'

Differential Revision: https://reviews.llvm.org/D147159